# check_repeat_nonrepeat Function Optimization