# Description

## Background
The `check_repeat_nonrepeat` function is used to identify edge case scenarios in longitudinal databases (those containing event arms) where REDCap users may make an instrument repeating _**and**_ nonrepeating in different events. This is a behavior that is not recommended and `REDCapTidieR` should not support. 

An example export would look something like below:

|record_id|redcap_event_name|redcap_repeat_instrument|redcap_repeat_instance|combination_variable|
|---------|-----------------|------------------------|----------------------|--------------------|
|1        |event_1          |NA                      |NA                    |A                   |
|2        |event_2          |combination             |1                     |B                   |
|3        |event_3          |combination             |2                     |C                   |

Where `combination_variable` contains data for rows where `redcap_repeat_*` is both `NA` and populated.

## Proposed Solution

When tested againt the Prodigy database, the previous method involved nested loops across all variables and their rows to check this behavior, resulting in `profvis` performance of 1740ms:
![image](https://user-images.githubusercontent.com/38384694/196510506-9fd33714-4282-4953-8a45-dd8829040a0d.png)

The new method creates dummy variables for each column and assigns values of "repeat" or "nonrepeat" per row. Then a `map`ping function determines if a dummy column has both "repeat" and "nonrepeat" identifiers and errors to return the name of the first column offender. The new performance has improved to 10ms:
![image](https://user-images.githubusercontent.com/38384694/196510778-abbe772e-4e0f-42f1-8a77-32b7c7fb5dd2.png)

# Proposed Changes 
List changes below in bullet format:

- Update the `check_repeat_nonrepeat` function
- Update `roxygen` documentation accordingly

### Issue Addressed
**Addresses some of the performance issues mentioned in #48, but more improvements are necessary.**

## Future Work

The performance improvement of `read_redcap_tidy` as a whole against the test Prodigy database improved from 23710ms to 18530ms, so there is still more improvement to go.

While the current method is better than the previous one, it may not be desirable to create a dataframe twice as large as the original database output. This could potentially be unappealing in exceptionally large database extractions.

### PR Checklist
Before submitting this PR, please check and verify below that the submission meets the below criteria:

- [x] New/revised functions have associated tests
- [NA] New tests that make API calls use `httptest::with_mock_api` and any new mocks were added to `tests/testthat/fixtures/create_httptest_mocks.R`
- [x] New/revised functions use appropriate naming conventions
- [x] New/revised functions don't repeat code
- [x] Code changes are less than **250** lines total
- [x] Issues linked to the PR using [GitHub's list of keywords](https://docs.github.com/en/issues/tracking-your-work-with-issues/linking-a-pull-request-to-an-issue)
- [x] The appropriate reviewer is assigned to the PR
- [x] The appropriate developers are assigned to the PR

# Code Review
This section to be used by the reviewer and developers during Code Review after PR submission

### Code Review Checklist

- [ ] I checked that new files follow naming conventions and are in the right place
- [ ] I checked that documentation is complete, clear, and without typos
- [ ] I added/edited comments to explain "why" not "how"
- [ ] I checked that all new variable and function names follow naming conventions
- [ ] I checked that new tests have been written for key business logic and/or bugs that this PR fixes
- [ ] I checked that new tests address important edge cases