@skadauke 2022-10-18
Yeah I think this is better! How about:

"Instrument detected that has both repeated and nonrepeated instances defined in the project: "
________________________________
From: Rich Hanna ***@***.***>
Sent: Tuesday, October 18, 2022 3:50 PM
To: CHOP-CGTDataOps/REDCapTidieR ***@***.***>
Cc: Stephan Kadauke ***@***.***>; Review requested ***@***.***>
Subject: [External]Re: [CHOP-CGTDataOps/REDCapTidieR] check_repeat_nonrepeat Function Optimization (PR #49)


@rsh52 commented on this pull request.

________________________________

In R/checks.R<https://nam10.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2FCHOP-CGTDataOps%2FREDCapTidieR%2Fpull%2F49%23discussion_r998645336&data=05%7C01%7Ckadaukes%40chop.edu%7C5310b421f19f456d217308dab14213d6%7Ca611241607b041a59bb1d146b575c975%7C0%7C0%7C638017194656872601%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C3000%7C%7C%7C&sdata=Zia9Un62hnHEebsCcgM66%2FRS6iPALjofACs3xAGCO00%3D&reserved=0>:

>        stop(
-        paste0("Instrument detected belonging to an instrument that is both repeating and nonrepeating: ",
-               db_data %>%
-                 select(.data$redcap_repeat_instrument, names(db_data)[i]) %>%
-                 drop_na(everything()) %>%
-                 slice(1) %>%
-                 pull(.data$redcap_repeat_instrument)
+        paste0("Instrument detected belonging to an instrument that is both


How about:
"Instrument detected with an assigned structure of both repeating and nonrepeating: "

—
Reply to this email directly, view it on GitHub<https://nam10.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2FCHOP-CGTDataOps%2FREDCapTidieR%2Fpull%2F49%23discussion_r998645336&data=05%7C01%7Ckadaukes%40chop.edu%7C5310b421f19f456d217308dab14213d6%7Ca611241607b041a59bb1d146b575c975%7C0%7C0%7C638017194657028835%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C3000%7C%7C%7C&sdata=X2WkpfRzbDSQtcdA4FS1c9aBEC0lqP%2FKSc44ZapaFu4%3D&reserved=0>, or unsubscribe<https://nam10.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Fnotifications%2Funsubscribe-auth%2FACTGHWU7ZHHA6Z2QWU7HHQTWD35SBANCNFSM6AAAAAARILIAO4&data=05%7C01%7Ckadaukes%40chop.edu%7C5310b421f19f456d217308dab14213d6%7Ca611241607b041a59bb1d146b575c975%7C0%7C0%7C638017194657028835%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C3000%7C%7C%7C&sdata=zBALYTbCHn3R6Ac1bVEx1MD6NTFFnHwGjQEWRptuYG0%3D&reserved=0>.
You are receiving this because your review was requested.Message ID: ***@***.***>

** This email originated from an EXTERNAL sender to CHOP. Proceed with caution when replying, opening attachments, or clicking links. Do not disclose your CHOP credentials, employee information, or protected health information to a potential hacker**.