@skadauke 2022-10-18
I think it's OK to say column #1 is safe since it's always the record ID and name can vary. However the additional columns I would enumerate by name.
________________________________
From: Rich Hanna ***@***.***>
Sent: Tuesday, October 18, 2022 3:54 PM
To: CHOP-CGTDataOps/REDCapTidieR ***@***.***>
Cc: Stephan Kadauke ***@***.***>; Review requested ***@***.***>
Subject: [External]Re: [CHOP-CGTDataOps/REDCapTidieR] check_repeat_nonrepeat Function Optimization (PR #49)


@rsh52 commented on this pull request.

________________________________

In R/checks.R<https://nam10.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2FCHOP-CGTDataOps%2FREDCapTidieR%2Fpull%2F49%23discussion_r998647951&data=05%7C01%7Ckadaukes%40chop.edu%7Cb1c7e775ecd44062e36d08dab1428e9e%7Ca611241607b041a59bb1d146b575c975%7C0%7C0%7C638017196660970123%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C3000%7C%7C%7C&sdata=jUr3L2OChXmEnBXQBh5DEzVqTolSuCQWzcnxxX9UgRc%3D&reserved=0>:

> -    repeating <- NA
-
-    for (j in seq_len(nrow(db_data))){
+  # This check function looks for potential repeat/nonrepeat behavior using the
+  # steps below:
+  # 1) Define standard columns that don't need checking and remove those from
+  #    analysis (i.e. "safe columns").
+  # 2) Create a dummy column for each remaining column and use case_when to
+  #    to assign whether the column demonstrates repeating, nonrepeating, or
+  #    indeterminant behavior per row. Indetemrinant would be the case for
+  #    data not yet filled out.
+  # 3) Use a mapping function to determine if any dummy columns contain both
+  #    "repeating" AND "nonrepeating" declarations, if so error out.
+
+  # Step (1)
+  safe_cols <- names(db_data)[1:4]


Ah that's a good point, I think I can correct this to look for the first 3 instead when redcap_event_name is missing i.e. is_longitudinal is T/F and reorder things.

—
Reply to this email directly, view it on GitHub<https://nam10.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2FCHOP-CGTDataOps%2FREDCapTidieR%2Fpull%2F49%23discussion_r998647951&data=05%7C01%7Ckadaukes%40chop.edu%7Cb1c7e775ecd44062e36d08dab1428e9e%7Ca611241607b041a59bb1d146b575c975%7C0%7C0%7C638017196661126347%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C3000%7C%7C%7C&sdata=F%2FI4KLATylqcYtjcTNh8ZHO7SaFEWSkgIpJgePFWJzo%3D&reserved=0>, or unsubscribe<https://nam10.safelinks.protection.outlook.com/?url=https%3A%2F%2Fgithub.com%2Fnotifications%2Funsubscribe-auth%2FACTGHWUK3PJUGJYGZE6QE2LWD3565ANCNFSM6AAAAAARILIAO4&data=05%7C01%7Ckadaukes%40chop.edu%7Cb1c7e775ecd44062e36d08dab1428e9e%7Ca611241607b041a59bb1d146b575c975%7C0%7C0%7C638017196661126347%7CUnknown%7CTWFpbGZsb3d8eyJWIjoiMC4wLjAwMDAiLCJQIjoiV2luMzIiLCJBTiI6Ik1haWwiLCJXVCI6Mn0%3D%7C3000%7C%7C%7C&sdata=EBoMjN1pyIHnPjwfqmFBzEtVaH2ryG7TCtO%2FgtooHvI%3D&reserved=0>.
You are receiving this because your review was requested.Message ID: ***@***.***>

** This email originated from an EXTERNAL sender to CHOP. Proceed with caution when replying, opening attachments, or clicking links. Do not disclose your CHOP credentials, employee information, or protected health information to a potential hacker**.